move expression out of exec subdirectory
