+++ cpp/examples/arrow/compute_register_example.cc
@@ -18,10 +18,10 @@
 #include <arrow/api.h>
 #include <arrow/compute/api.h>
 #include <arrow/compute/exec/exec_plan.h>
-#include <arrow/compute/exec/expression.h>
 #include <arrow/compute/exec/options.h>
 #include <arrow/util/async_generator.h>
 #include <arrow/util/future.h>
+#include "arrow/compute/expression.h"
 
 #include <cstdlib>
 #include <iostream>
