+++ cpp/src/arrow/compute/exec/CMakeLists.txt
@@ -17,12 +17,11 @@
 
 arrow_install_all_headers("arrow/compute/exec")
 
-add_arrow_compute_test(expression_test
+add_arrow_compute_test(subtree_test
                        REQUIRE_ALL_KERNELS
                        PREFIX
                        "arrow-compute"
                        SOURCES
-                       expression_test.cc
                        subtree_test.cc)
 
 add_arrow_compute_test(plan_test
