+++ cpp/src/arrow/compute/exec/fetch_node.cc
@@ -21,11 +21,11 @@
 #include "arrow/compute/exec.h"
 #include "arrow/compute/exec/accumulation_queue.h"
 #include "arrow/compute/exec/exec_plan.h"
-#include "arrow/compute/exec/expression.h"
 #include "arrow/compute/exec/map_node.h"
 #include "arrow/compute/exec/options.h"
 #include "arrow/compute/exec/query_context.h"
 #include "arrow/compute/exec/util.h"
+#include "arrow/compute/expression.h"
 #include "arrow/datum.h"
 #include "arrow/result.h"
 #include "arrow/util/checked_cast.h"
