+++ cpp/src/arrow/compute/exec/project_benchmark.cc
@@ -23,10 +23,10 @@
 #include "arrow/compute/cast.h"
 #include "arrow/compute/exec.h"
 #include "arrow/compute/exec/benchmark_util.h"
-#include "arrow/compute/exec/expression.h"
 #include "arrow/compute/exec/options.h"
 #include "arrow/compute/exec/task_util.h"
 #include "arrow/compute/exec/test_util.h"
+#include "arrow/compute/expression.h"
 #include "arrow/dataset/partition.h"
 #include "arrow/testing/future_util.h"
 #include "arrow/testing/generator.h"
