+++ cpp/src/arrow/compute/exec/sink_node.cc
@@ -25,12 +25,12 @@
 #include "arrow/compute/exec.h"
 #include "arrow/compute/exec/accumulation_queue.h"
 #include "arrow/compute/exec/exec_plan.h"
-#include "arrow/compute/exec/expression.h"
 #include "arrow/compute/exec/options.h"
 #include "arrow/compute/exec/order_by_impl.h"
 #include "arrow/compute/exec/query_context.h"
 #include "arrow/compute/exec/util.h"
 #include "arrow/compute/exec_internal.h"
+#include "arrow/compute/expression.h"
 #include "arrow/datum.h"
 #include "arrow/result.h"
 #include "arrow/table.h"
