+++ cpp/src/arrow/compute/exec/source_node.cc
@@ -21,11 +21,11 @@
 
 #include "arrow/compute/exec.h"
 #include "arrow/compute/exec/exec_plan.h"
-#include "arrow/compute/exec/expression.h"
 #include "arrow/compute/exec/options.h"
 #include "arrow/compute/exec/query_context.h"
 #include "arrow/compute/exec/util.h"
 #include "arrow/compute/exec_internal.h"
+#include "arrow/compute/expression.h"
 #include "arrow/datum.h"
 #include "arrow/io/util_internal.h"
 #include "arrow/result.h"
