+++ cpp/src/arrow/compute/exec/type_fwd.h
@@ -17,50 +17,20 @@
 
 #pragma once
 
-#include "arrow/util/visibility.h"
+#include "arrow/compute/type_fwd.h"
 
 namespace arrow {
 
-struct Datum;
-struct TypeHolder;
-
 namespace compute {
 
-class Function;
-class FunctionExecutor;
-class FunctionOptions;
-class FunctionRegistry;
+class ExecNode;
+class ExecPlan;
+class ExecNodeOptions;
+class ExecFactoryRegistry;
+class QueryContext;
+struct QueryOptions;
 
-/// \brief Return the process-global function registry.
-// Defined in registry.cc
-ARROW_EXPORT FunctionRegistry* GetFunctionRegistry();
-
-class CastOptions;
-
-struct ExecBatch;
-class ExecContext;
-class KernelContext;
-
-struct Kernel;
-struct ScalarKernel;
-struct ScalarAggregateKernel;
-struct VectorKernel;
-
-struct KernelState;
-
-struct Declaration;
-class Expression;
-// class ExecNode;
-// class ExecPlan;
-// class ExecNodeOptions;
-// class ExecFactoryRegistry;
-// class QueryContext;
-// struct QueryOptions;
-//
-// class SinkNodeConsumer;
-
-ARROW_EXPORT ExecContext* default_exec_context();
-ARROW_EXPORT ExecContext* threaded_exec_context();
+class SinkNodeConsumer;
 
 }  // namespace compute
 }  // namespace arrow
