+++ cpp/src/arrow/dataset/dataset.h
@@ -26,7 +26,7 @@
 #include <utility>
 #include <vector>
 
-#include "arrow/compute/exec/expression.h"
+#include "arrow/compute/expression.h"
 #include "arrow/dataset/type_fwd.h"
 #include "arrow/dataset/visibility.h"
 #include "arrow/util/async_generator_fwd.h"
