+++ cpp/src/arrow/dataset/partition.h
@@ -28,7 +28,7 @@
 #include <utility>
 #include <vector>
 
-#include "arrow/compute/exec/expression.h"
+#include "arrow/compute/expression.h"
 #include "arrow/dataset/type_fwd.h"
 #include "arrow/dataset/visibility.h"
 #include "arrow/util/compare.h"
