+++ cpp/src/arrow/dataset/scanner_test.cc
@@ -28,8 +28,8 @@
 #include "arrow/compute/api_vector.h"
 #include "arrow/compute/cast.h"
 #include "arrow/compute/exec/exec_plan.h"
-#include "arrow/compute/exec/expression_internal.h"
 #include "arrow/compute/exec/test_util.h"
+#include "arrow/compute/expression_internal.h"
 #include "arrow/dataset/dataset_internal.h"
 #include "arrow/dataset/plan.h"
 #include "arrow/dataset/test_util_internal.h"
