+++ cpp/src/skyhook/cls/cls_skyhook.cc
@@ -18,7 +18,7 @@
 
 #include <memory>
 
-#include "arrow/compute/exec/expression.h"
+#include "arrow/compute/expression.h"
 #include "arrow/dataset/dataset.h"
 #include "arrow/dataset/file_ipc.h"
 #include "arrow/dataset/file_parquet.h"
