+++ cpp/src/skyhook/protocol/skyhook_protocol_test.cc
@@ -16,7 +16,7 @@
 // under the License.
 #include "skyhook/protocol/skyhook_protocol.h"
 
-#include "arrow/compute/exec/expression.h"
+#include "arrow/compute/expression.h"
 #include "arrow/dataset/test_util_internal.h"
 #include "arrow/table.h"
 #include "arrow/testing/gtest_util.h"
